requirements.txt
selenium
pyperclip
openai
python-dotenv
blinker==1.7.0
webdriver-manager
selenium-wire
httpx